Extract classes TaskList and Ui from Sigma.java
